refactor(readr): update import type path
