+++ packages/readr/graphql/fragments/post.ts
@@ -1,5 +1,19 @@
 import gql from 'graphql-tag'
 
+import type { GenericPost } from '~/types/common'
+
+export type Post = Pick<
+  GenericPost,
+  | 'id'
+  | 'slug'
+  | 'style'
+  | 'title'
+  | 'heroImage'
+  | 'ogImage'
+  | 'publishTime'
+  | 'readingTime'
+>
+
 import { photoFragment } from './photo'
 export const postFragment = gql`
   fragment PostFields on Post {
